+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Character editing features:
 
 - Added undo/redo keyboard shortcuts.
 - Added custom character how to.
+- Fixed #712, adding pixel eraser.
 
 ### Fixed